+++ packages/domain-post/src/index.ts
@@ -1,1 +1,8 @@
-export {};
+import { registerPostMappings } from './models/mapper.js';
+
+export * from './entities/index.js';
+export * from './models/index.js';
+export * from './services/index.js';
+export * from './repositories/index.js';
+
+registerPostMappings();